Bugfix in "Hiller Lab Views" to include supertracks (implemented by Pradeep Kiruvale). Detailed description is in the log.
**HillerLabView Project Description**

Hiller lab view project is about, to save the existing track view settings in the database and loading, deleting views on the fly.

Technical details.

Files which have undergone changes are

1.hgTracks.c
2.hgTracks.h
3.Makefile

and added two new files

1.hillerLabView.c
2.hillerLabView.h

Some Implementation details to know before understanding how it works.
Genome browser UI provides a text box to enter a view name, store view button, load view button and a dropdown box with list of existing views and delete view button.
All views are stored in database table \u201cHillerLabViews\u201d in hgcentraltest. This table contains two columns one for view name and the settings (big string blob). A track list containing all the tracks supported by the particular species will be prepared upfront for that species, once it is selected. The cart will act as an agent of for all view operations.

Work flow:-

*Save view operation*
To save a view user has to enter a view name in the text box and press save-view button. Once the user presses button, exiting configuration of all the tracks will be captured through the cart hash structure. Then all the track related entries in the cart will be fetched according to the existing track list for that particular species, the variable name and value pair will be separated by tab and put into string array. Each entry (name-value pair) will be separated by a newline chartecter.
Once the extraction of data from the cart is done, the prepared string will be stored in a database (hgcentraltest) with view-name as a primary key.
Successive save view operation will add a new row to the table. And dropdown list entries will be updated.

*Load View Operation*
All the views stored in the database will have an entry in the dropdown list of views in the browser. So to load a view, user has to select a view from the dropdown list and press on load view button, once the user presses the button the particular view will be fetched from the database and the cart will be filled with these name-value pair, comparing against the track list for that particular
Species. Once the filling of cart is done, the browser will display loaded view.

*Delete view operation*
To delete a view, user has to select the view from the dropdown list and should press delete view button once it\u2019s pressed, the particular row related to the view in database will be removed, and dropdown list entries will be updated.

NOTE:-We won't store custom track data (includes marked region as well as custom tracks)
Views can be loaded across species; Means the view stored for a species1 can be loaded in species2. But only the related track data for the current species will be loaded and displayed.